clang-tidy fixes

+++ native/app/mru.h
@@ -10,7 +10,7 @@ extern "C" {
 int init_mru(const char *mru_path);
 void deinit_mru();
 int queue_mru_search(uv_loop_t *loop, const char *value, const char *mru_path);
-int write_mru(const char *mru_path, const char *path);
+size_t write_mru(const char *mru_path, const char *path);
 int is_mru_search_ongoing();
 void init_mru_mutex();
 void deinit_mru_mutex();
